rename/change username into login

+++ database/crud/db_users.py
@@ -2,15 +2,14 @@ from sqlalchemy.orm import Session
 from database.models import DbUsers
 from schemas.schemas import CreateNewUser
 from database.hash import Hash
-from fastapi import HTTPException, status
 from typing import Type
 
 
 def create_user(db: Session, request: CreateNewUser) -> DbUsers:
     """Create a new record in DbUsers with given data."""
-    username = request.username
+    login = request.login.lower()
     hash_password = Hash().bcrypt_pass(request.password)
-    user = DbUsers(username=username,
+    user = DbUsers(login=login,
                    password=hash_password,
                    )
     db.add(user)
@@ -25,39 +24,36 @@ def get_all_users(db: Session) -> list[Type[DbUsers]]:
     return all_users
 
 
-def get_user(user_id: int | None, username: str | None, db: Session) -> Type[DbUsers] | None:
-    """Find record in DbUsers with given username or user_id. Returns found record."""
+def get_user(user_id: int | None, login: str | None, db: Session) -> Type[DbUsers] | None:
+    """Find record in DbUsers with given login or user_id. Returns found record."""
+    get_login = login.lower()
     if user_id:
         user = db.query(DbUsers).filter_by(user_id=user_id).first()
-        if user is None:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                                detail=f"There's no User with id: {user_id}",
-                                )
         return user
-    user = db.query(DbUsers).filter_by(username=username).first()
-    if user is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                            detail=f"There's no User with username: {username}",
-                            )
+    user = db.query(DbUsers).filter_by(login=get_login).first()
     return user
 
 
-def update_user(user_id: int | None, username: str | None, db: Session, request: CreateNewUser) -> bool:
-    """Find record in DbUsers with given user_id or username. Update record with new data."""
-    new_username = request.username
+def update_user(user_id: int | None, login: str | None, db: Session, request: CreateNewUser) -> bool:
+    """Find record in DbUsers with given user_id or login. Update record with new data."""
+    new_login = request.login.lower()
     new_hash_password = Hash().bcrypt_pass(request.password)
     if user_id:
-        user = get_user(username=None, user_id=user_id, db=db)
+        user = get_user(login=None, user_id=user_id, db=db)
+        if user is None:
+            return False
         user.update(
-            {DbUsers.username: new_username,
+            {DbUsers.login: new_login,
              DbUsers.password: new_hash_password,
              }
         )
         db.commit()
         return True
-    user = get_user(username=username, user_id=None, db=db)
+    user = get_user(login=login, user_id=None, db=db)
+    if user is None:
+        return False
     user.update(
-        {DbUsers.username: new_username,
+        {DbUsers.login: new_login,
          DbUsers.password: new_hash_password,
          }
     )
@@ -65,14 +61,19 @@ def update_user(user_id: int | None, username: str | None, db: Session, request:
     return True
 
 
-def delete_user(user_id: int | None, username: str | None, db: Session) -> bool:
-    """Delete record from DbUsers with given user_id or username."""
+def delete_user(user_id: int | None, login: str | None, db: Session) -> bool:
+    """Delete record from DbUsers with given user_id or login."""
+    del_login = login.lower()
     if user_id:
-        user = get_user(username=None, user_id=user_id, db=db)
+        user = get_user(login=None, user_id=user_id, db=db)
+        if user is None:
+            return False
         db.delete(user)
         db.commit()
         return True
-    user = get_user(username=username, user_id=None, db=db)
+    user = get_user(login=del_login, user_id=None, db=db)
+    if user is None:
+        return False
     db.delete(user)
     db.commit()
     return True
